requirements.txt
python-telegram-bot==20.3
groq==0.4.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.0